refactor(S12P11A404-123): 이미지 관련 Service - 로컬 저장 형식에서 BLOB 저장 형식으로 변경

+++ server/src/main/java/com/kiosk/server/store/service/impl/ImageDeleteServiceImpl.java
@@ -1,14 +1,11 @@
 package com.kiosk.server.store.service.impl;
 
+import com.kiosk.server.common.exception.custom.UnauthorizedException;
 import com.kiosk.server.store.domain.ImageRepository;
 import com.kiosk.server.store.domain.Images;
 import com.kiosk.server.store.service.ImageDeleteService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 
 @Service
 @RequiredArgsConstructor
@@ -17,27 +14,22 @@ public class ImageDeleteServiceImpl implements ImageDeleteService {
     private final ImageRepository imageRepository;
 
     @Override
-    public void doService(int imageId) {
+    public void doService(long userId, int imageId) {
 
-        if (imageId == 0){
+        if (imageId == 0) {
             return; // 이미지 없는 경우 삭제하지 않음
         }
 
         // 기존 이미지 정보 조회
         Images exImage = imageRepository.findImageById(imageId);
-        if (exImage == null){
+        if (exImage == null) {
             return;
         }
-
-        // 파일 삭제
-        try{
-            Path imagePath = Paths.get(exImage.getImagePath());
-            Files.deleteIfExists(imagePath);
-        } catch (Exception e){
-            throw new RuntimeException("기존 이미지 삭제 실패");
+        if (exImage.getUserId() != userId) {
+            throw new UnauthorizedException("Mismatched userId");
         }
 
-        // DB에서 삭제
+        // DB에서 이미지 삭제
         imageRepository.deleteImageById(imageId);
     }
 }
